# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Running Trainer App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trainees:
    get:
      operationId: listTrainees
      tags: [trainees]
      summary: List all trainees
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: planType
          in: query
          schema:
            type: string
            enum: [free, paid]
        - name: planned
          in: query
          schema:
            type: boolean
        - name: hasBalance
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of trainees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trainee"
    post:
      operationId: createTrainee
      tags: [trainees]
      summary: Create a trainee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTraineeInput"
      responses:
        "201":
          description: Created trainee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trainee"
  /trainees/{id}:
    get:
      operationId: getTrainee
      tags: [trainees]
      summary: Get a trainee by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trainee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trainee"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateTrainee
      tags: [trainees]
      summary: Update a trainee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTraineeInput"
      responses:
        "200":
          description: Updated trainee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trainee"
    delete:
      operationId: deleteTrainee
      tags: [trainees]
      summary: Delete a trainee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /trainees/{id}/transactions:
    get:
      operationId: listTraineeTransactions
      tags: [transactions]
      summary: List transactions for a trainee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Add a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionInput"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /trainees/{id}/transactions/{txId}:
    delete:
      operationId: deleteTransaction
      tags: [transactions]
      summary: Delete a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: txId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ──────────────── SEGMENTS ────────────────
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List all events
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create an event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventInput"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}:
    put:
      operationId: updateEvent
      tags: [events]
      summary: Update an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventInput"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ──────────────── DASHBOARD ────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ─── TRAINEE ───
    Trainee:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        city:
          type: string
        birthdate:
          type: string
          format: date
        runsPerWeek:
          type: integer
        hrZone4:
          type: integer
        hrZone5a:
          type: integer
        hrZone5c:
          type: integer
        targetHr:
          type: integer
        targetSpeedFrom:
          type: string
        targetSpeedTo:
          type: string
        testDate:
          type: string
          format: date
        lactateThresholdHr:
          type: integer
        preferredPayment:
          type: string
          enum: [cash, credit_card]
        planType:
          type: string
          enum: [free, paid]
        planFinishDate:
          type: string
          format: date
        monthlyFee:
          type: number
        balanceDue:
          type: number
        hasActivityThisWeek:
          type: boolean
        isPlannedThisWeek:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - planType
        - createdAt
    CreateTraineeInput:
      type: object
      properties:
        name:
          type: string
        city:
          type: string
        birthdate:
          type: string
          format: date
        runsPerWeek:
          type: integer
        hrZone4:
          type: integer
        hrZone5a:
          type: integer
        hrZone5c:
          type: integer
        targetHr:
          type: integer
        targetSpeedFrom:
          type: string
        targetSpeedTo:
          type: string
        testDate:
          type: string
          format: date
        lactateThresholdHr:
          type: integer
        preferredPayment:
          type: string
          enum: [cash, credit_card]
        planType:
          type: string
          enum: [free, paid]
        planFinishDate:
          type: string
          format: date
        monthlyFee:
          type: number
      required:
        - name
        - planType

    # ─── CONTRACT ───
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        traineeId:
          type: integer
        date:
          type: string
          format: date
        amount:
          type: number
        activityMonth:
          type: string
        method:
          type: string
          enum: [cash, credit_card, online]
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - traineeId
        - date
        - amount
        - activityMonth
        - createdAt
    CreateTransactionInput:
      type: object
      properties:
        date:
          type: string
          format: date
        amount:
          type: number
        activityMonth:
          type: string
        method:
          type: string
          enum: [cash, credit_card, online]
        notes:
          type: string
      required:
        - date
        - amount
        - activityMonth
...
    Event:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        date:
          type: string
          format: date
        location:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - date
        - isActive
        - createdAt
    CreateEventInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        date:
          type: string
          format: date
        location:
          type: string
        isActive:
          type: boolean
      required:
        - name
        - date

    # ─── DASHBOARD ───
...