# lib/api-zod/src/generated/types/createEventInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Running Trainer App API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateEventInput {
  name: string;
  description?: string;
  date: Date;
  location?: string;
  isActive?: boolean;
}