# lib/api-zod/src/generated/types/createTransactionInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Running Trainer App API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateTransactionInputMethod } from "./createTransactionInputMethod";

export interface CreateTransactionInput {
  date: Date;
  amount: number;
  activityMonth: string;
  method?: CreateTransactionInputMethod;
  notes?: string;
}